Use size_t instead of int for stack ptr

+++ Zend/zend_ptr_stack.h
@@ -23,7 +23,7 @@
 #include "zend_alloc.h"
 
 typedef struct _zend_ptr_stack {
-	int top, max;
+	size_t top, max;
 	void **elements;
 	void **top_element;
 	bool persistent;
@@ -41,7 +41,7 @@ ZEND_API void zend_ptr_stack_destroy(zend_ptr_stack *stack);
 ZEND_API void zend_ptr_stack_apply(zend_ptr_stack *stack, void (*func)(void *));
 ZEND_API void zend_ptr_stack_reverse_apply(zend_ptr_stack *stack, void (*func)(void *));
 ZEND_API void zend_ptr_stack_clean(zend_ptr_stack *stack, void (*func)(void *), bool free_elements);
-ZEND_API int zend_ptr_stack_num_elements(zend_ptr_stack *stack);
+ZEND_API size_t zend_ptr_stack_num_elements(zend_ptr_stack *stack);
 END_EXTERN_C()
 
 #define ZEND_PTR_STACK_RESIZE_IF_NEEDED(stack, count)		\
